Add property-based tests and README badges

+++ easy_hash/src/ordered_float.rs
@@ -1,10 +1,10 @@
-use crate::{split_u64, type_salt, EasyHash};
+use crate::{EasyHash, split_u64, type_salt};
 
 impl EasyHash for ordered_float::OrderedFloat<f32> {
     const TYPE_SALT: u32 = type_salt::<Self>();
     fn ehash(&self) -> u64 {
         let mut checksum = fletcher::Fletcher64::new();
-        checksum.update(&[Self::TYPE_SALT, self.0.to_bits() as u32]);
+        checksum.update(&[Self::TYPE_SALT, self.0.to_bits()]);
         checksum.value()
     }
 }
@@ -13,7 +13,7 @@ impl EasyHash for ordered_float::OrderedFloat<f64> {
     fn ehash(&self) -> u64 {
         let mut checksum = fletcher::Fletcher64::new();
         checksum.update(&[Self::TYPE_SALT]);
-        let bits = self.to_bits() as u64;
+        let bits = self.to_bits();
         checksum.update(&split_u64(bits));
         checksum.value()
     }
@@ -23,7 +23,7 @@ impl EasyHash for ordered_float::NotNan<f32> {
     const TYPE_SALT: u32 = type_salt::<Self>();
     fn ehash(&self) -> u64 {
         let mut checksum = fletcher::Fletcher64::new();
-        checksum.update(&[Self::TYPE_SALT, self.to_bits() as u32]);
+        checksum.update(&[Self::TYPE_SALT, self.to_bits()]);
         checksum.value()
     }
 }
@@ -33,7 +33,7 @@ impl EasyHash for ordered_float::NotNan<f64> {
     fn ehash(&self) -> u64 {
         let mut checksum = fletcher::Fletcher64::new();
         checksum.update(&[Self::TYPE_SALT]);
-        let bits = self.to_bits() as u64;
+        let bits = self.to_bits();
         checksum.update(&split_u64(bits));
         checksum.value()
     }
